Improving CSS class name
Previously `navigation-controls-container` was both an element ID
as well as a CSS class. This was confusing, especially because the
element with that ID was not one of the elements with that CSS
class.

The CSS class is used on elements that contain a section of the
navigation controls container, so the word "section" seems more
appropriate to use in the class name than "container". This also
differentiates it from the element ID.