Swapping order of "Zoom In"/"Zoom Out" buttons
This new ordering seems possibly more logical, because if zooming
was instead handled via a slider the leftmost position would
probably correspond to the most zoomed out position, and the
rightmost position the most zoomed in position. This ordering also
matches the ordering of the horizontally laid out "-"/"+" buttons
in Apple Maps. (Google Maps lays out the corresponding buttons
vertically so there isn't necessarily an apples-to-apples
comparison possible there).