Only allowing main mouse button to affect gestures
Previously, the `mousedown` and `mouseup` event handlers responded
to any mouse button being pressed, and this could lead to
problems. For example, if the middle mouse button was clicked
while continuously pressing the main mouse button an in-progress
gesture would stop, which wasn't the intended behavior. (At least,
I believe this is what happened, but I don't have a mouse like
this to test with). After this commit, only the main (i.e. usually
left) mouse button can start or stop a gesture.

A particular benefit is that this prevents gesture handling from
going into a bad state when a context menu is shown via clicking
on the main `<canvas>` element with the secondary (i.e. usually
right) mouse button. Previously, after the context menu was closed
moving the mouse would cause gestures to occur (such as a pan in
manual mode or look-around in flythrough mode) even if no mouse
buttons were pressed. Clicking the main mouse button on the
`<canvas>` element would restore things back to a good state.

However, after this commit gesture handling can still go into a
bad state if a context menu is shown via other means, such as by
pressing the control key on the keyboard while pressing the main
mouse button in macOS. (At least, in Safari 17.4 and Chrome
122.0.6261.129, but not Firefox 123.0.1 because it inaccurately
indicates the secondary mouse button was pressed in this
situation). Fixing context menu issues like this will be done
separately.