Fixing "invisible city" flythrough bugs
Previously, there were multiple ways make the entire scene briefly
disappear in flythrough mode. This happened whenever the main
`<canvas>` element received a `mouseup` event without having first
received a corresponding `mousedown` event. For example, one way
to trigger this was to start a flythrough, make sure the mouse was
somewhere in the navigation controls box, press the main mouse
button, move the mouse onto the main `<canvas>` element, and
release the main mouse button.

In flythrough mode a `mousedown` event causes the camera's
direction to be locked in the vehicle's current direction, and a
`mouseup` event causes an animation from the locked direction to
the vehicle's updated direction. Previously, an isolated `mouseup`
event resulted in a faulty animation in which the starting X/Y
direction angles were `NaN`. This caused the camera's X/Y
direction angles to be `NaN` for the duration of the animation.
(Thus the city was't actually invisible, the camera just wasn't
looking at it).

This commit fixes these bugs by preventing the `mouseup` event
handler from having any effect unless preceded by a `mousedown`
event. This seems like good behavior in general, even for manual
mode where I'm not aware of the previous behavior causing any
problems.

It would probably be a good idea to also fix the underlaying
animation logic in the future as well, but at least in the short
term not fixing it makes it easier to tell if there are similar
bugs with event handling still lurking.