Standardizing way of checking mouse button status
Previously, there were three different ways in use to determine if
the main mouse button was pressed (one of which only did so
indirectly). After this commit there is only a single consistent
way used.

A side benefit is that this allows the `previousTouches()`
getter functions to be removed from the gesture processor objects,
since they are no longer used anywhere.