Simplifying `mousemove` event handling
This commit makes `mousemove` events on the main `<canvas>`
element handled in the same way as `mouseup` events. That is,
dynamically added when a `mousedown` event occurs, and removed
after a `mouseup` event. This is simpler than the previous
implementation that had to manually keep track of whether the main
mouse button was pressed, or whether the mouse had left or
re-entered the `<canvas>` element.