Preventing gesture problems due to context menus
Commit 71ca11b343fbd8e19f914f9d6c39788325b4ad8d prevented gesture
handling from going into a bad state when a context menu was shown
by clicking the secondary mouse button on the main `<canvas>`
element. This commit fixes additional ways that showing a context
menu could cause problems with gesture handling.

For example, one such way in Safari and Chrome on macOS was
showing a context menu by pressing the main mouse button on the
`<canvas>` element while pressing the control key on the keyboard.
Previously, after the context menu was closed moving the mouse
would cause gestures to occur even if the main mouse button wasn't
pressed. (This particular issue did not happen in Firefox because
as of v124.0.1 it inaccurately indicates the secondary mouse
button was pressed in this scenario).